Move ca commands to its own package.
